Move kernel check to be a os-agnostic thing
